requirements.txt
Flask==3.1.3
requests==2.32.5
python-dotenv==1.2.1
icalendar==7.0.2
gunicorn==25.1.0